observer implementer

+++ src/main/java/org/location/models/Chauffeur.java
@@ -1,13 +1,12 @@
 package org.location.models;
 
-import javax.persistence.*;
-
+import  javax.persistence.*;
 @Entity
 @Table(name = "chauffeurs")
 public class Chauffeur {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(nullable = false)
     private String nom;
@@ -22,8 +21,8 @@ public class Chauffeur {
         this.dispo = dispo;
     }
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
 
     public String getNom() { return nom; }
     public void setNom(String nom) { this.nom = nom; }
